Refactor and remove `utils.js`
