+++ src/avs/charts/OperatorsTabLineChart.jsx
@@ -1,5 +1,6 @@
 import { AxisBottom, AxisRight } from '@visx/axis';
 import { Circle, LinePath } from '@visx/shape';
+import { getGrowthPercentage, reduceState } from '../../shared/helpers';
 import { scaleLinear, scaleUtc } from '@visx/scale';
 import { Tab, Tabs } from '@nextui-org/react';
 import { useCallback, useEffect, useMemo } from 'react';
@@ -7,11 +8,9 @@ import { useTooltip, useTooltipInPortal } from '@visx/tooltip';
 import { bisector } from '@visx/vendor/d3-array';
 import { curveMonotoneX } from '@visx/curve';
 import { formatNumber } from '../../shared/formatters';
-import { getGrowthPercentage } from '../../utils';
 import { GridRows } from '@visx/grid';
 import { Group } from '@visx/group';
 import { localPoint } from '@visx/event';
-import { reduceState } from '../../shared/helpers';
 import { tabs } from '../../shared/slots';
 import { useMutativeReducer } from 'use-mutative';
 import { useTailwindBreakpoint } from '../../shared/useTailwindBreakpoint';
