+++ src/avs/charts/TVLTabLineChart.jsx
@@ -1,17 +1,16 @@
 import { AxisBottom, AxisRight } from '@visx/axis';
 import { Circle, LinePath } from '@visx/shape';
 import { formatETH, formatNumber, formatUSD } from '../../shared/formatters';
+import { getGrowthPercentage, reduceState } from '../../shared/helpers';
 import { scaleLinear, scaleUtc } from '@visx/scale';
 import { Tab, Tabs } from '@nextui-org/react';
 import { useCallback, useEffect, useMemo } from 'react';
 import { useTooltip, useTooltipInPortal } from '@visx/tooltip';
 import { bisector } from '@visx/vendor/d3-array';
 import { curveMonotoneX } from '@visx/curve';
-import { getGrowthPercentage } from '../../utils';
 import { GridRows } from '@visx/grid';
 import { Group } from '@visx/group';
 import { localPoint } from '@visx/event';
-import { reduceState } from '../../shared/helpers';
 import { tabs } from '../../shared/slots';
 import { useMutativeReducer } from 'use-mutative';
 import { useTailwindBreakpoint } from '../../shared/useTailwindBreakpoint';
